Fix intermittent closing of unrelated handles after closeKey.
After a handle has been closed explicitly, it must not be closed again by the finalizer, because the handle value may have been reused for unrelated things in the meantime.

The handle must be invalidated by closeKey, and because N-API "external" values are immutable, that requires boxing the HKEY value in another dynamically managed allocation.