Fix intermittent closing of unrelated handles after closeKey.
After a handle has been closed explicitly, it must not be closed again by the finalizer, because the handle value may have been reused for unrelated things in the meantime.

The handle would have to be invalidated by closeKey, which is not straightforward to implement with N-API "external" values because they are immutable. For now, instead remove the finalizer and make it the user's responsibility not to leak keys, restoring the behavior from before the N-API rewrite.